Removed 'git' options from the getfiles script. Added new option to pull down the latest versions of files

+++ tools/getfiles.py
@@ -28,9 +28,6 @@ LOG.setLevel(logging.INFO)
 ## ==============================================
 ## DEFAULT CONFIGURATION
 ## ==============================================
-GIT_REPO = "git://github.com/apavlo/h-store-files.git"
-GIT_BRANCH = "master"
-
 SVN_REPO = "https://database.cs.brown.edu/svn/hstore-files/"
 SVN_OPTS = "--non-interactive --trust-server-cert"
 
@@ -42,10 +39,8 @@ if __name__ == '__main__':
     aparser.add_argument('path', help='Installation path')
     aparser.add_argument('--svn-repo', default=SVN_REPO, help='SVN repository')
     aparser.add_argument('--svn-options', default=SVN_OPTS, help='SVN checkout options')
-    aparser.add_argument('--git', action='store_true', help='Use git instead of svn')
-    aparser.add_argument('--git-repo', default=GIT_REPO, help='Git repository')
-    aparser.add_argument('--git-branch', default=GIT_BRANCH, help='Git branch')
     aparser.add_argument('--overwrite', action='store_true', help='Overwrite existing directory')
+    aparser.add_argument('--update', action='store_true', help='Pull down the latest version of files')
     aparser.add_argument('--copy', default=None, help='Copy from existing local copy')
     aparser.add_argument('--symlink', default=None, help='Create a symlink for the directory')
     args = vars(aparser.parse_args())
@@ -69,23 +64,28 @@ if __name__ == '__main__':
     
     if copy:
         cmd = "cp -rvl %(copy)s %(path)s" % args
-    elif args['git']:
-        cmd = "git clone --branch %(git_branch)s %(git_repo)s %(path)s" % args
     else:
         cmd = "svn %(svn_options)s checkout %(svn_repo)s %(path)s" % args
 
-    if os.path.exists(args['path']):
-        if not args['overwrite']:
-            LOG.info("Installation directory '%s' already exists. Not overwriting" % os.path.realpath(args['path']))
-            sys.exit(0)
+    if os.path.exists(args['path']) and args['overwrite']:
         LOG.warn("Deleting directory '%s' and reinstalling" % args['path'])
         shutil.rmtree(args['path'])
-
-    # Bombs away!
-    LOG.info(cmd)
-    subprocess.check_call(cmd, shell=True)
-    
-    if not args['symlink'] is None and not os.path.islink(args['symlink']):
-        LOG.info("Creating symlink from %(path)s to %(symlink)s" % args)
-        os.symlink(args["path"], args["symlink"])
+        
+    if not os.path.exists(args['path']):
+        # Bombs away!
+        LOG.info(cmd)
+        subprocess.check_call(cmd, shell=True)
+        if not args['symlink'] is None and not os.path.islink(args['symlink']):
+            LOG.info("Creating symlink from %(path)s to %(symlink)s" % args)
+            os.symlink(args["path"], args["symlink"])
+    else:
+        LOG.info("Installation directory '%s' already exists. Not overwriting" % os.path.realpath(args['path']))
+        
+    ## Update
+    if args['update']:
+        os.chdir(args["path"])
+        cmd = "svn %(svn_options)s update" % args
+        LOG.info(cmd)
+        subprocess.check_call(cmd, shell=True)
+        
 ## MAIN
